// UI/app/types/music.d.ts
export interface Track {
  trackId: number;
  title: string;
  duration: number;
  imageUrl: string;
  audioUrl: string;
  artistName: string;
  genre: string;
  viewCount: number;
  isVisible: boolean;
}

export interface Playlist {
  playlistId: number;
  name: string;
  coverImages: string[];
  tracks: Track[];
}



export interface DetailViewData {
  type: "playlist" | "artist";
  title: string;
  coverImages: string[];
  tracks: Track[];
}

export type SelectedItem = DetailViewData;

// User
export interface User {
  userId: number;
  username: string;
  email: string;
  role: string;
  actStatus: string;
  isActive: boolean;
  avatar: string;
}

export interface UserProfileData {
  firstName: string;
  lastName: string;
  gender: string;
  dateOfBirth: string;
  phone: string;
  address: string;
}